+++ ContentPatcher/release-notes.md
@@ -5,7 +5,9 @@
 * Updated for Stardew Valley 1.4.
 * Added support for data model assets in Stardew Valley 1.4.
 * Added support for tokens in field keys.
+* Improved subkey validation.
 * Fixed incorrect error text when dynamic/config tokens conflict.
+* Removed support for the subkey syntax in condition values (where you could do `"Relationship": "Abigail:Married"` instead of `"Relationship:Abigail": "Married"`). This only affects one content pack on Nexus.
 
 ## 1.6.5
 * Fixed `EditData` allowing field values containing `/` (which is the field delimiter).